Adding Optimization Algorithm
Here we are using genetic algorithm in our problem to estimate the maximum profit values over a year. 
The generation of genomes is based on the value of MwMedMes = 50 for 12 months, meaning that the maximum value of 600 has been assigned (so far) to limit the generation of individuals under this condition, since 600/12 = 50. The fitness values are set by a two variable complex function with restrictions. these restrictions are set on the crossover function where -5 <= x1 <= 5 and -5 <= x2 <= 5.